Do some refactoring and start adding tests for shapes.
