+++ glue_ar/gltf_utils.py
@@ -1,4 +1,6 @@
-from typing import List
+import operator
+import struct
+from typing import Callable, Iterable, List, Optional, Type, TypeVar, Union
 
 from gltflib import Material, PBRMetallicRoughness
 
@@ -8,6 +10,10 @@ __all__ = [
     "bring_into_clip",
     "offset_triangles",
     "create_material_for_color",
+    "add_points_to_bytearray",
+    "add_triangles_to_bytearray",
+    "index_mins",
+    "index_maxes",
 ]
 
 
@@ -50,3 +56,37 @@ def create_material_for_color(
             ),
             alphaMode="BLEND"
     )
+
+
+def add_points_to_bytearray(arr: bytearray, points: Iterable[Iterable[Union[int, float]]]):
+    for point in points:
+        for coordinate in point:
+            arr.extend(struct.pack('f', coordinate))
+
+
+def add_triangles_to_bytearray(arr: bytearray, triangles: Iterable[Iterable[int]]):
+    for triangle in triangles:
+        for index in triangle:
+            arr.extend(struct.pack('I', index))
+
+
+T = TypeVar("T", bound=Union[int, float])
+
+
+def index_extrema(items: List[List[T]],
+                  extremum: Callable[[T, T], T],
+                  previous: Optional[List[List[T]]] = None,
+                  type: Type[T] = float) -> List[List[T]]:
+    size = len(items[0])
+    extrema = [type(extremum([operator.itemgetter(i)(item) for item in items])) for i in range(size)]
+    if previous is not None:
+        extrema = [extremum(x, p) for x, p in zip(extrema, previous)]
+    return extrema
+
+
+def index_mins(items, previous=None, type: Type[T] = float) -> List[List[T]]:
+    return index_extrema(items, extremum=min, type=type, previous=previous)
+
+
+def index_maxes(items, previous=None, type: Type[T] = float) -> List[List[T]]:
+    return index_extrema(items, extremum=max, type=type, previous=previous)
