+++ tox.ini
@@ -1,3 +1,17 @@
+[tox]
+envlist = py{38,39,310,311,312}-test
+
+[testenv]
+passenv =
+    DISPLAY
+    HOME
+changedir =
+    test: .tmp/{envname}
+commands =
+    test: pip freeze
+    test: pytest --pyargs glue_ar --cov glue_ar {posargs}
+
+
 [testenv:codestyle]
 deps = flake8
 skipsdist = true
